Refactor setMonto ingreso y egreso

+++ gastanding/Egreso.h
@@ -21,7 +21,7 @@ public:
 	int getMedioPago() { return _medioPago; };
 	string getConcepto() { return string(_concepto); };
 	bool getEstado() { return _estado; };
-	void setMonto(int valor) { _monto = valor; };
+	void setMonto(float valor) { _monto = valor; };
 	void setConcepto(string valor) { strcpy_s(_concepto, valor.c_str()); };
 	void setCategoria(int valor) { _categoria = valor; };
 	void setMedioPago(int valor) { _medioPago = valor; };
